Fixed up various things around CubeEntity, so that materials works and its possible to change material in Editor, and so that material is stored in savefile, among other things that I don't remember I changed.

+++ Components/Engine/entities/EntityComponentSystem.h
@@ -11,7 +11,7 @@ namespace voxie {
 
     class Handle {
     public:
-        using Identity = int;
+        using Identity = u_int64_t;
         explicit Handle(Identity);
 
         static Handle MakeEntity();
